+++ bloom_nofos/nofos/forms.py
@@ -47,6 +47,7 @@ NofoTaglineForm = create_nofo_form_class(["tagline"])
 NofoThemeForm = create_nofo_form_class(["theme"])
 NofoCoverForm = create_nofo_form_class(["cover"])
 NofoStatusForm = create_nofo_form_class(["status"])
+NofoIconPathForm = create_nofo_form_class(["icon_path"])
 
 
 # this one needs a custom field and a custom widget so don't use the factory function