+++ CHANGELOG.md
@@ -10,6 +10,7 @@ The format is based on Keep a Changelog, and this project adheres to Semantic Ve
 
 - Add 'print' button to NOFO edit view
 - Text only cover page for HRSA light theme
+- Add new NOFO field: icon_path
 
 ### Changed